fix: main 함수 중복 문제 해결 및 공통 테마 형식으로 수정 (#21)
